+++ crates/maybenot/src/constants.rs
@@ -2,7 +2,11 @@
 
 /// The highest possible version of a [`Machine`](crate::Machine) supported by
 /// this framework.
-pub const VERSION: u8 = 2;
+pub const VERSION: u8 = 3;
+
+/// Legacy version 2 (bincode-based serialization).
+#[cfg(feature = "legacy-v02")]
+pub const VERSION_LEGACY: u8 = 2;
 
 /// The maximum size of a decompressed encoded [`Machine`](crate::Machine) in
 /// bytes. Set to 1MB. This is a soft limit and can be increased if necessary.